Henter og bruker oppgave for inntektsrapportering.
