+++ src/main/kotlin/no/nav/brukerdialog/ytelse/ungdomsytelse/api/UngdomsytelseService.kt
@@ -9,12 +9,16 @@ import no.nav.brukerdialog.integrasjon.ungdeltakelseopplyser.UngDeltakelseOpplys
 import no.nav.brukerdialog.metrikk.MetrikkService
 import no.nav.brukerdialog.utils.MDCUtil
 import no.nav.brukerdialog.utils.TokenUtils.personIdent
+import no.nav.brukerdialog.ytelse.ungdomsytelse.api.domene.inntektsrapportering.OppgittInntektForPeriode
+import no.nav.brukerdialog.ytelse.ungdomsytelse.api.domene.inntektsrapportering.UngPeriode
 import no.nav.brukerdialog.ytelse.ungdomsytelse.api.domene.inntektsrapportering.UngdomsytelseInntektsrapportering
+import no.nav.brukerdialog.ytelse.ungdomsytelse.api.domene.inntektsrapportering.UngdomsytelseInntektsrapporteringInnsending
 import no.nav.brukerdialog.ytelse.ungdomsytelse.api.domene.oppgavebekreftelse.UngdomsytelseOppgavebekreftelseInnsending
 import no.nav.brukerdialog.ytelse.ungdomsytelse.api.domene.oppgavebekreftelse.UngdomsytelseOppgavebekreftelse
 import no.nav.brukerdialog.ytelse.ungdomsytelse.api.domene.soknad.Barn
 import no.nav.brukerdialog.ytelse.ungdomsytelse.api.domene.soknad.Ungdomsytelsesøknad
 import no.nav.security.token.support.spring.SpringTokenValidationContextHolder
+import no.nav.ung.deltakelseopplyser.kontrakt.oppgave.felles.InntektsrapporteringOppgavetypeDataDTO
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Service
 import java.util.*
@@ -48,13 +52,33 @@ class UngdomsytelseService(
     }
 
     suspend fun inntektrapportering(rapportetInntekt: UngdomsytelseInntektsrapportering, gitSha: String) {
-        val metadata = MetaInfo(correlationId = MDCUtil.callIdOrNew(), soknadDialogCommitSha = gitSha)
-        val cacheKey = "${springTokenValidationContextHolder.personIdent()}_${rapportetInntekt.ytelse()}"
+        val rapporterInntektOppgave = ungDeltakelseOpplyserService.hentOppgaveForDeltakelse(
+            UUID.fromString(rapportetInntekt.oppgaveReferanse)
+        )
 
-        logger.info(formaterStatuslogging(rapportetInntekt.ytelse(), rapportetInntekt.innsendingId(), "mottatt."))
+        val inntektsrapporteringOppgaveData = (rapporterInntektOppgave.oppgavetypeData as? InntektsrapporteringOppgavetypeDataDTO
+            ?: throw IllegalStateException("OppgavetypeData er ikke av type InntektsrapporteringOppgavetypeDataDTO"))
+
+        val inntektsrapporteringInnsending = UngdomsytelseInntektsrapporteringInnsending(
+            oppgaveReferanse = rapportetInntekt.oppgaveReferanse,
+            mottatt = rapportetInntekt.mottatt,
+            oppgittInntektForPeriode = OppgittInntektForPeriode(
+                arbeidstakerOgFrilansInntekt = rapportetInntekt.oppgittInntekt.arbeidstakerOgFrilansInntekt,
+                periodeForInntekt = UngPeriode(
+                    fraOgMed = inntektsrapporteringOppgaveData.fraOgMed,
+                    tilOgMed = inntektsrapporteringOppgaveData.tilOgMed
+                )
+            ),
+            harBekreftetInntekt = rapportetInntekt.harBekreftetInntekt,
+        )
+
+        val metadata = MetaInfo(correlationId = MDCUtil.callIdOrNew(), soknadDialogCommitSha = gitSha)
+        val cacheKey = "${springTokenValidationContextHolder.personIdent()}_${inntektsrapporteringInnsending.ytelse()}"
+
+        logger.info(formaterStatuslogging(inntektsrapporteringInnsending.ytelse(), inntektsrapporteringInnsending.innsendingId(), "mottatt."))
         duplikatInnsendingSjekker.forsikreIkkeDuplikatInnsending(cacheKey)
-        innsendingService.registrer(rapportetInntekt, metadata)
-        metrikkService.registrerMottattInnsending(rapportetInntekt.ytelse())
+        innsendingService.registrer(inntektsrapporteringInnsending, metadata)
+        metrikkService.registrerMottattInnsending(inntektsrapporteringInnsending.ytelse())
     }
 
     suspend fun oppgavebekreftelse(oppgavebekreftelse: UngdomsytelseOppgavebekreftelse, gitSha: String) {
